inital work on using tidyselect for feature columns
This demonstrates the use of tidyselect and piping for definition of feature columns. Note that I haven't yet done all of the feature column types (just 5 of them). This is partly because the crossed features don't fit as neatly into tidyselect (note however that I'm pretty sure we can make them work).