don't automatically deduce buckets from data
Since the data argument passed to `feature_columns` need only be an sample/exemplar of the data, I think its dangerous to deduce the number of buckets from it. This change requires explicit specification of num_buckets, and also allows for passing a simple character vector of column names (rather than a dataset) to feature_columns.